Add files via upload
A simple recommendation system that suggests items to
users based on their preferences. You can use techniques like
collaborative filtering or content-based filtering to recommend
movies, books, or products to users.
This is more useful to the users to analysis the data from the suggestions.
It is the one of the prior way to get the information about the movies.